Update after working together
Changed predictors and tried to change the outcome to lat/long, but we need to have a combined column to do so and I am not quite sure how we want to do that.  I would have liked to combine them as a coordinate pair, but that will not work because the value needs to be numerical.  For the time being, I just set the outcome variable as latitude so the code will at least give a linear regression.  I also eliminated all data, not from the US.