Add world cloud visual, fixed linear regression
I added a world cloud visual at the bottom of the code, I needed to install some stuff to make it work so be aware of that if you want to run it yourself, otherwise, I added a photo of it to the slides.  I did the subtraction thing to combine the lat and long columns, and updated all of our models to use this new variable.  As a result, the accuracy of the model is not great, but it is something.